add DbMessageRepository

+++ src/main/java/com/plux/distribution/domain/message/MessageId.java
@@ -1,5 +1,5 @@
 package com.plux.distribution.domain.message;
 
-public record MessageId(Integer value) {
+public record MessageId(Long value) {
 
 }
